refactor(constants): Combined home and page constants
Since home constants uses page constants web giving an error with page constants not being defined.
