+++ src/modules/compass/Compass.js
@@ -18,9 +18,9 @@ import {useDispatch, useSelector} from 'react-redux';
 import {isEmpty, roundToDecimalPlaces} from '../../shared/Helpers';
 import deviceSound from '../../utils/sounds/sound';
 import {formStyles} from '../form';
-import {MODAL_KEYS} from '../home/home.constants';
 import {setModalVisible} from '../home/home.slice';
 import useMeasurementsHook from '../measurements/useMeasurements';
+import {MODAL_KEYS} from '../page/page.constants';
 import {COMPASS_TOGGLE_BUTTONS} from './compass.constants';
 import {setCompassMeasurements} from './compass.slice';
 import compassStyles from './compass.styles';
