+++ src/modules/measurements/AddMeasurementModal.js
@@ -16,10 +16,10 @@ import Compass from '../compass/Compass';
 import {setCompassMeasurementTypes} from '../compass/compass.slice';
 import compassStyles from '../compass/compass.styles';
 import {Form, formStyles, useFormHook} from '../form';
-import {MODAL_KEYS} from '../home/home.constants';
 import {setModalValues, setModalVisible} from '../home/home.slice';
 import overlayStyles from '../home/overlay.styles';
 import useLocationHook from '../maps/useLocation';
+import {MODAL_KEYS} from '../page/page.constants';
 import {updatedModifiedTimestampsBySpotsIds} from '../project/projects.slice';
 import {editedSpotProperties, setSelectedAttributes} from '../spots/spots.slice';
 import Templates from '../templates/Templates';
