+++ src/modules/notes/Notes.js
@@ -7,10 +7,9 @@ import {useDispatch, useSelector} from 'react-redux';
 import {isEmpty} from '../../shared/Helpers';
 import SaveButton from '../../shared/SaveButton';
 import uiStyles from '../../shared/ui/ui.styles';
-import {MODAL_KEYS} from '../home/home.constants';
 import useLocationHook from '../maps/useLocation';
 import {setNotebookPageVisible} from '../notebook-panel/notebook.slice';
-import {PAGE_KEYS, PRIMARY_PAGES} from '../page/page.constants';
+import {MODAL_KEYS, PAGE_KEYS, PRIMARY_PAGES} from '../page/page.constants';
 import ReturnToOverviewButton from '../page/ui/ReturnToOverviewButton';
 import {updatedModifiedTimestampsBySpotsIds} from '../project/projects.slice';
 import {editedSpotProperties, setSelectedSpotNotesTimestamp} from '../spots/spots.slice';
