+++ src/modules/sketch/Sketch.js
@@ -5,10 +5,10 @@ import {useNavigation} from '@react-navigation/native';
 import RNSketchCanvas from '@StraboSpot/react-native-sketch-canvas';
 import {useDispatch, useSelector} from 'react-redux';
 
-import {MODAL_KEYS} from '../home/home.constants';
 import {setModalVisible} from '../home/home.slice';
 import useImagesHook from '../images/useImages';
 import useLocationHook from '../maps/useLocation';
+import {MODAL_KEYS} from '../page/page.constants';
 import {updatedModifiedTimestampsBySpotsIds} from '../project/projects.slice';
 import {editedSpotImages} from '../spots/spots.slice';
 import styles from './sketch.styles';
