+++ src/shared/ui/modal/ModalHeader.js
@@ -4,9 +4,8 @@ import {Text, View} from 'react-native';
 import {Button} from 'react-native-elements';
 import {useDispatch, useSelector} from 'react-redux';
 
-import {MODALS} from '../../../modules/home/home.constants';
 import {setModalVisible} from '../../../modules/home/home.slice';
-import {NOTEBOOK_PAGES, PAGE_KEYS} from '../../../modules/page/page.constants';
+import {MODALS, NOTEBOOK_PAGES, PAGE_KEYS} from '../../../modules/page/page.constants';
 import * as themes from '../../styles.constants';
 import modalStyle from './modal.style';
